+++ __fixtures__/v-next/outputv2/cosmos/evidence/v1beta1/tx.ts
@@ -259,7 +259,7 @@ export const MsgSubmitEvidenceResponse = {
 };
 export const Evidence_InterfaceDecoder = (input: BinaryReader | Uint8Array): Any => {
   const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
-  const data = Any.decode(reader, reader.uint32(), true);
+  const data = Any.decode(reader, reader.uint32());
   switch (data.typeUrl) {
     default:
       return data;