+++ __fixtures__/v-next/outputv2/cosmos/feegrant/v1beta1/tx.ts
@@ -478,14 +478,14 @@ export const MsgRevokeAllowanceResponse = {
 };
 export const FeeAllowanceI_InterfaceDecoder = (input: BinaryReader | Uint8Array): BasicAllowance | PeriodicAllowance | AllowedMsgAllowance | Any => {
   const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
-  const data = Any.decode(reader, reader.uint32(), true);
+  const data = Any.decode(reader, reader.uint32());
   switch (data.typeUrl) {
     case "/cosmos.feegrant.v1beta1.BasicAllowance":
-      return BasicAllowance.decode(data.value, undefined, true);
+      return BasicAllowance.decode(data.value);
     case "/cosmos.feegrant.v1beta1.PeriodicAllowance":
-      return PeriodicAllowance.decode(data.value, undefined, true);
+      return PeriodicAllowance.decode(data.value);
     case "/cosmos.feegrant.v1beta1.AllowedMsgAllowance":
-      return AllowedMsgAllowance.decode(data.value, undefined, true);
+      return AllowedMsgAllowance.decode(data.value);
     default:
       return data;
   }