+++ __fixtures__/v-next/outputv2/cosmos/staking/v1beta1/tx.ts
@@ -1400,7 +1400,7 @@ export const MsgUndelegateResponse = {
 };
 export const Cosmos_cryptoPubKey_InterfaceDecoder = (input: BinaryReader | Uint8Array): Any => {
   const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
-  const data = Any.decode(reader, reader.uint32(), true);
+  const data = Any.decode(reader, reader.uint32());
   switch (data.typeUrl) {
     default:
       return data;