+++ __fixtures__/v-next/outputv2/cosmwasm/wasm/v1/types.ts
@@ -1373,7 +1373,7 @@ export const Model = {
 };
 export const ContractInfoExtension_InterfaceDecoder = (input: BinaryReader | Uint8Array): Any => {
   const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
-  const data = Any.decode(reader, reader.uint32(), true);
+  const data = Any.decode(reader, reader.uint32());
   switch (data.typeUrl) {
     default:
       return data;