+++ packages/ast/src/clients/rpc/extension/rpc.extension.test.ts
@@ -3,7 +3,7 @@ import { getNestedProto } from '@cosmology/utils'
 import { defaultTelescopeOptions, ProtoService } from '@cosmology/types';
 import { expectCode, getTestProtoStore } from '../../../../test-utils';
 import { GenericParseContext } from '../../../encoding';
-import { createRpcClientImpl, createRpcQueryExtension } from './extension';
+import { createRpcClientImpl, createRpcMsgExtension, createRpcQueryExtension } from './extension';
 const store = getTestProtoStore();
 store.traverseAll();
 
@@ -13,6 +13,14 @@ it('createRpcQueryExtension', () => {
     const service: ProtoService = getNestedProto(res).Query;
     const context = new GenericParseContext(ref, store, defaultTelescopeOptions);
     expectCode(createRpcQueryExtension(context, service))
+});
+
+it('createRpcMsgExtension', () => {
+  const ref = store.findProto('cosmos/auth/v1beta1/query.proto');
+  const res = traverse(store, ref);
+  const service: ProtoService = getNestedProto(res).Query;
+  const context = new GenericParseContext(ref, store, defaultTelescopeOptions);
+  expectCode(createRpcMsgExtension(context, 'getCosmosTxRpc'))
 });
 
 it('createRpcClientImpl', () => {