+++ packages/parser/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.1](https://github.com/cosmology-tech/telescope/compare/@cosmology/proto-parser@1.4.0...@cosmology/proto-parser@1.4.1) (2024-01-08)
+
+**Note:** Version bump only for package @cosmology/proto-parser
+
 # [1.4.0](https://github.com/cosmology-tech/telescope/compare/@cosmology/proto-parser@1.3.2...@cosmology/proto-parser@1.4.0) (2023-12-26)
 
 **Note:** Version bump only for package @cosmology/proto-parser