+++ packages/telescope/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.3](https://github.com/cosmology-tech/telescope/compare/@cosmology/telescope@1.4.2...@cosmology/telescope@1.4.3) (2024-01-08)
+
+**Note:** Version bump only for package @cosmology/telescope
+
+## [1.4.2](https://github.com/cosmology-tech/telescope/compare/@cosmology/telescope@1.4.1...@cosmology/telescope@1.4.2) (2024-01-08)
+
+**Note:** Version bump only for package @cosmology/telescope
+
+## [1.4.1](https://github.com/cosmology-tech/telescope/compare/@cosmology/telescope@1.4.0...@cosmology/telescope@1.4.1) (2024-01-02)
+
+**Note:** Version bump only for package @cosmology/telescope
+
 # [1.4.0](https://github.com/cosmology-tech/telescope/compare/@cosmology/telescope@1.3.3...@cosmology/telescope@1.4.0) (2023-12-26)
 
 **Note:** Version bump only for package @cosmology/telescope